GETP-193 rename: infra 패키지에서 mail과 storage 추출
