+++ src/main/java/es/princip/getp/domain/auth/infra/EmailVerificationCodeSender.java
@@ -3,8 +3,8 @@ package es.princip.getp.domain.auth.infra;
 import es.princip.getp.domain.auth.application.VerificationCodeSender;
 import es.princip.getp.domain.auth.exception.FailedVerificationCodeSendingException;
 import es.princip.getp.domain.member.command.domain.model.Email;
-import es.princip.getp.infra.mail.EmailSender;
-import es.princip.getp.infra.mail.command.SendMailCommand;
+import es.princip.getp.mail.EmailSender;
+import es.princip.getp.mail.command.SendMailCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.MailException;
 import org.springframework.stereotype.Service;
