+++ src/main/java/es/princip/getp/domain/project/command/infra/EmailMeetingSender.java
@@ -5,8 +5,8 @@ import es.princip.getp.domain.project.command.application.MeetingSender;
 import es.princip.getp.domain.project.command.domain.Project;
 import es.princip.getp.domain.project.command.domain.ProjectMeeting;
 import es.princip.getp.domain.project.exception.FailedMeetingSendingException;
-import es.princip.getp.infra.mail.EmailSender;
-import es.princip.getp.infra.mail.command.SendMailCommand;
+import es.princip.getp.mail.EmailSender;
+import es.princip.getp.mail.command.SendMailCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.MailException;
 import org.springframework.stereotype.Service;
