+++ src/test/java/es/princip/getp/infra/storage/presentation/FileStorageControllerTest.java
@@ -1,8 +1,9 @@
 package es.princip.getp.infra.storage.presentation;
 
 import es.princip.getp.infra.annotation.WithCustomMockUser;
-import es.princip.getp.infra.storage.application.FileUploadService;
+import es.princip.getp.storage.application.FileUploadService;
 import es.princip.getp.infra.support.ControllerTest;
+import es.princip.getp.storage.presentation.FileStorageController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
